refactor debug page
