+++ site/catalog/checkboxes.js
@@ -13,7 +13,7 @@ const colors = [
   'purple'
 ];
 
-class CheckboxesRadios extends React.Component {
+class Checkboxes extends React.Component {
   render() {
     const checkmark = (
       <svg
@@ -49,39 +49,12 @@ class CheckboxesRadios extends React.Component {
       );
     });
 
-    const radios = colors.map((color, i) => {
-      let radioClass = 'radio mr6';
-      let id = 'radio';
-      if (color !== null) {
-        radioClass += ` radio--${color}`;
-        id += `-${color}`;
-      }
-
-      return (
-        <div key={i} className='mb12'>
-          <label className='radio-container mr12'>
-            <input id={id} name='radios' value={id} type='radio' />
-            <div className={radioClass}></div>
-            Radio
-          </label>
-          <label className='radio-container'>
-            <input checked disabled id={id} name={id} value={id} type='radio' />
-            <div className={radioClass}></div>
-            Radio
-          </label>
-        </div>
-      );
-    });
-
     return (
       <div>
-        <h1 className='txt-headline mb18'>
-          Checkboxes & Radios
-        </h1>
-
-        <h2 className='subheadline mb18 mt18'>
+        <h2 className='border-b border--2 border--gray-faint pb6 mt72 mb24 txt-l txt-uppercase txt-bold'>
           Checkboxes
         </h2>
+
         <div className='mb12'>
           <input id='checkbox-unstyled' type='checkbox' />
           <label for='checkbox-unstyled'>Unstyled</label>
@@ -89,18 +62,9 @@ class CheckboxesRadios extends React.Component {
 
         {checkboxes}
 
-        <h2 className='subheadline mb18 mt18'>
-          Checkboxes
-        </h2>
-        <div className='mb12'>
-          <input id='radio-unstyled' type='radio' />
-          <label for='radio-unstyled'>Radio</label>
-        </div>
-        {radios}
-
       </div>
     );
   }
 }
 
-export { CheckboxesRadios };
+export { Checkboxes };
