+++ site/catalog/selects.js
@@ -1,0 +1,83 @@
+import React from 'react';
+
+const colors = [
+  null,
+  'gray',
+  'pink',
+  'red',
+  'orange',
+  'yellow',
+  'green',
+  'teal',
+  'blue',
+  'purple'
+];
+
+
+class Selects extends React.Component {
+  render() {
+    return (
+      <div>
+        <h2 className='border-b border--2 border--gray-faint pb6 mt72 mb24 txt-l txt-uppercase txt-bold'>
+          Selects
+        </h2>
+
+        {colors.map((color) => {
+          let selectClass = 'select';
+          let selectStrokeClass = 'select select--stroke';
+          const selectContainerClass = 'select-container';
+          if (color !== null) {
+            selectClass += ` select--${color}`;
+            selectStrokeClass += ` select--stroke-${color}`;
+          }
+          return (
+            <div key={color} className='mb12'>
+              <div className='inline-block mr12'>
+                <div className={selectContainerClass}>
+                  <select className={selectClass}>
+                    <option>firstoption</option>
+                    <option>two</option>
+                    <option>three</option>
+                  </select>
+                  <div className='select-arrow'></div>
+                </div>
+              </div>
+              <div className='inline-block mr12'>
+                <div className={selectContainerClass}>
+                  <select className={selectClass} disabled>
+                    <option>firstoption</option>
+                    <option>two</option>
+                    <option>three</option>
+                  </select>
+                  <div className='select-arrow'></div>
+                </div>
+              </div>
+              <div className='inline-block mr12'>
+                <div className={selectContainerClass}>
+                  <select className={selectStrokeClass}>
+                    <option>firstoption</option>
+                    <option>two</option>
+                    <option>three</option>
+                  </select>
+                  <div className='select-arrow'></div>
+                </div>
+              </div>
+              <div className='inline-block mr12'>
+                <div className={selectContainerClass}>
+                  <select className={`${selectClass} select--s`}>
+                    <option>firstoption</option>
+                    <option>two</option>
+                    <option>three</option>
+                  </select>
+                  <div className='select-arrow'></div>
+                </div>
+              </div>
+            </div>
+          );
+        })}
+      </div>
+    );
+  }
+}
+
+export { Selects };
