--- site/debug/spinners.js
@@ -1,23 +1,0 @@
-import React from 'react';
-
-class Spinners extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1 className='txt-headline mb18'>
-          Spinners
-        </h1>
-
-        <div>
-
-          <div className="loading p12"></div>
-          <div className="loading loading--s p12"></div>
-          <div className="loading loading--dark bg-gray-dark p12"></div>
-
-        </div>
-      </div>
-    );
-  }
-}
-
-export { Spinners };
